Revert nesting changes to ConsoleDrivers
