+++ Terminal.Gui/ConsoleDrivers/CursesDriver/binding.cs
@@ -48,9 +48,23 @@ namespace Unix.Terminal {
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
 	public partial class Curses {
+		//[StructLayout (LayoutKind.Sequential)]
+		//public struct winsize {
+		//	public ushort ws_row;
+		//	public ushort ws_col;
+		//	public ushort ws_xpixel;   /* unused */
+		//	public ushort ws_ypixel;   /* unused */
+		//};
+
+		[StructLayout (LayoutKind.Sequential)]
+		public struct MouseEvent {
+			public short ID;
+			public int X, Y, Z;
+			public Event ButtonState;
+		}
 
 		static int lines, cols;
-		static CursesWindow main_window;
+		static Window main_window;
 		static IntPtr curses_handle, curscr_ptr, lines_ptr, cols_ptr;
 
 		// If true, uses the DllImport into "ncurses", otherwise "libncursesw.so.5"
@@ -83,7 +97,7 @@ namespace Unix.Terminal {
 			cols_ptr = get_ptr ("COLS");
 		}
 
-		static public CursesWindow initscr ()
+		static public Window initscr ()
 		{
 			setlocale (LC_ALL, "");
 			FindNCurses ();
@@ -91,7 +105,7 @@ namespace Unix.Terminal {
 			// Prevents the terminal from being locked after exiting.
 			reset_shell_mode ();
 
-			main_window = new CursesWindow (methods.initscr ());
+			main_window = new Window (methods.initscr ());
 			try {
 				console_sharp_get_dims (out lines, out cols);
 			} catch (DllNotFoundException) {
@@ -296,8 +310,8 @@ namespace Unix.Terminal {
 		static public int init_pair (short pair, short f, short b) => methods.init_pair (pair, f, b);
 		static public int use_default_colors () => methods.use_default_colors ();
 		static public int COLOR_PAIRS () => methods.COLOR_PAIRS ();
-		static public uint getmouse (out CursesMouseEvent ev) => methods.getmouse (out ev);
-		static public uint ungetmouse (ref CursesMouseEvent ev) => methods.ungetmouse (ref ev);
+		static public uint getmouse (out MouseEvent ev) => methods.getmouse (out ev);
+		static public uint ungetmouse (ref MouseEvent ev) => methods.ungetmouse (ref ev);
 		static public int mouseinterval (int interval) => methods.mouseinterval (interval);
 		static public bool is_term_resized (int lines, int columns) => methods.is_term_resized (lines, columns);
 		static public int resize_term (int lines, int columns) => methods.resize_term (lines, columns);
@@ -372,8 +386,8 @@ namespace Unix.Terminal {
 		public delegate int init_pair (short pair, short f, short b);
 		public delegate int use_default_colors ();
 		public delegate int COLOR_PAIRS ();
-		public delegate uint getmouse (out CursesMouseEvent ev);
-		public delegate uint ungetmouse (ref CursesMouseEvent ev);
+		public delegate uint getmouse (out Curses.MouseEvent ev);
+		public delegate uint ungetmouse (ref Curses.MouseEvent ev);
 		public delegate int mouseinterval (int interval);
 		public delegate IntPtr mousemask (IntPtr newmask, out IntPtr oldMask);
 		public delegate bool is_term_resized (int lines, int columns);
